Refactoring & Tuning some components
